requirements.txt
fastapi
uvicorn
newspaper3k
nltk
lxml